+++ src/main/resources/META-INF/services/io.camunda.connector.api.outbound.OutboundConnectorFunction
@@ -1,1 +1,1 @@
-io.camunda.connector.sap.ODataConnector
+io.camunda.connector.sap.odata.ODataConnector
